final api

+++ src/modules/wallet/useCases/deleteWallet/DeleteWalletUseCase.ts
@@ -1,3 +1,4 @@
+import { IActionsRepository } from "@modules/actions/repositories/IActionsRepository";
 import { IWalletRepository } from "@modules/wallet/repositories/IWalletRepository";
 import { inject, injectable } from "tsyringe";
 
@@ -8,6 +9,8 @@ class DeleteWalletUseCase {
     constructor(
         @inject("WalletRepository")
         private walletRepository: IWalletRepository,
+        @inject("ActionsRepository")
+        private actionsRepository: IActionsRepository,
     ) {}
 
     async execute(id: string, user_id: string): Promise<void> {
@@ -15,6 +18,16 @@ class DeleteWalletUseCase {
 
         if (wallet.user_id !== user_id) {
             throw new Error("User not permision!")
+        }
+
+        if (!wallet) {
+            throw new Error("Wallet not exists!")
+        }
+
+        const actions = await this.actionsRepository.findByWalletId(id);
+
+        for(const action of actions) {
+            await this.actionsRepository.delete(action.id);
         }
 
         await this.walletRepository.delete(id);
